Changed depth type from float to double

+++ src/Rasterizer.cpp
@@ -38,7 +38,7 @@ const double Rasterizer::calculateDepth(const Triangle3D& triangle_world, const 
   double u, v, w;
   triangle_screen.calculateBarycentricCoords(u, v, w, pixel_screen);
   // Interpolate Z in 3D using inverse function and barycentric coordinates in 2D
-  const float depth = 1 / 
+  const double depth = 1 / 
     (
       (1 / triangle_camera.v1.position.z) * u + 
       (1 / triangle_camera.v2.position.z) * v +
